Attempt to simplify code using `create_from_yaml`
Unfortunately, this isn't quite what we were looking for. We actually
want to `apply`, not `create`, because with `create` (the imperative
method) we need to deal with cleanup of the previous object, or patch
it. With `apply` (the declarative method), we specify what we want and
kubernetes takes care of the rest.