Update docs for running the CLI in dev
Will add GHA to publish `ogdc-runner` image to ghcr, which will remove the need
for the local build. Also need to support toggling "dev mode" on/off so that we
can specify `image_pull_policy="Never"` only in a dev context when we want to
use a locally built image. If this is not set, Argo will try to pull the image
from Dockerhub and fail.