WIP very wip refactor around using volumes instead of artifacts
It isn't easy to reference artifacts between workflows. We might want to utilzie
artifacts internally within workflows, but at the edges (inputs and outputs)
volumes seem like the best interface.

The downside here is that we have to manage isolation between workflow runs and
steps. our code is responsible for making sure that the paths are unique and
separated by steps...