Resolve typecheck errors
Some of the errors around traversing nodes and checking outputs is a bit
confusing.  I think the way we have the workflow setup means that the relevant
attrs will be present. May want to consider more robust error checking (or maybe
wrap all of it in try/except...) down the road.