Cleanup test workflow after execution
Will revisit this. The `submit_ogdc_recipe` function may end up submitting more
than one workflow that we want to preserve (or cleanup) in the future, which
would mean that we can't just return the name of a single workflow. Maybe we end
up having a result object that contains references to all workflows executed for
a given recipe.