Rename "simple" recipe to "shell"
This reflects the fact that commands are executed as shell commands against the
provided docker image (`ogdc-runner` by default).

We could be more explicit by calling this a "shell_script" or "sh_script", but
for conciseness, I thnk this works. I would not anticipate `ogdc-runner`
supporting any other kinds of shells, or scripts.