Pydantic config models: seprate `meta.yaml` config
When writing documentation for #78, I realized that the `RecipeConfig` I was
linking to had one piece of data that we do not expect users to directly set in
the `meta.yaml` - the `recipe_directory`. This is given at the CLI and utlized
internally. I imagine there might be other bits of "recipe config" that are also
not expressed in the meta.yaml, so separating out a `RecipeMeta`  model that is
specific to the `meta.yaml` itself is handy, and makes things a little more
clear wrt documentation.