Stage provided OGDC recipe directory locally to temp dir
This ensures that if we pass an fsspec-compatible string representing a remote
recipe directory (via GitHub) via the CLI, that directory gets copied to a
local staging directory.

This lets all other code that interacts with the recipe directory to access it
directly on disk, without needing to know anything about the source file system
protocol.

Right now the staging directory is a tempdir that gets cleaned up at the end of
program execution. I think that when we move to an architecture with a service
in front of argo deployed to k8s, that we might want to stage the recpie to a
PVC for archival/provenance. It can live in the same directory that workflow
outputs live.