Combine runner image w/ service image
For now, I think it is cleanest to have both responsibilities taken by a single
image. We ancipate possibly making use of <TODO> soon, which would require this
sort of setup anyway.

If the need for other curated runners comes about, then setting up new repos for
them might be the best option.