Remove `image_pull_policy="Never"` in local env
Skaffold now handles ensuring the locally-built ogdc-runner image is deployed to
the cluster, which was the reason this was originally setup - to ensure that
only the locally & manually built docker image was used. This is OBE.

This was breaking the integration tests for viz workflow, which use other
images (e.g., `alpine`) and are not managed by skaffold/placed on the cluster manually.